geoalchemy/requirements.txt
cx_Oracle
python-dotenv
geopandas